refactor: Move the sync FeatureStoreClientWrapper into the FDv2 data system
FeatureStoreClientWrapper lived in fdv2_common.py, which is meant for logic
shared between the sync FDv2 and the async AsyncFDv2 (via _FDv2Base). This
wrapper is sync-only — only fdv2.py uses it; the async side has its own
wrapper in async_fdv2.py. Move it into fdv2.py, next to its sole user, and
drop the now-unused imports and __all__ entry from fdv2_common.py. This
mirrors the earlier FDv1 wrapper move (5948b78). Pure relocation, no
behavior change.
